refactor(#4016): extract the search-value subsystem; 35/35 in the addressed lanes
The #3102 LOC ratchet caught this correctly: both files touched are god-files
already at their cap and the first version grew them +255 / +55. The gate's own
instruction is "add code to the subsystem module, not the barrel/driver", so the
§22.1.3 search-value dispatch now lives in `src/codegen/string-search-value.ts` —
the admissibility predicates, the `RegExpCreate(ToString(v))` lowering, and the
coerced search/match/split entry points.

The split is meaningful rather than cosmetic: the new module owns the DECISION
(is the spec's plain-ToString path the whole of the semantics here?) while
`regexp-standalone.ts` keeps the engine plumbing it calls into. Growth drops to
+19 / +19 (-93% / -65%); what remains is the seam only — one import per file, a
few `export` keywords, the `regexpOverride` option on the two shared emitters,
and a two-line delegation at each of the three call sites. `loc-budget-allow`
covers that residual, not the logic.

This also FIXES the 4 residuals rather than merely moving them, and refutes the
explanation the previous revision gave for them. They were reported as
"`any`-typed arguments — the conservative boundary of `wellKnownSymbolMemberOf`",
a plausible story that matched the intended design and so read as a finding. The
real cause was a scripted `str.replace()` that silently did not match
(whitespace), printed `ok`, and left the search/match gates on the narrower
`isStaticallyUndefinedExpr` while split got the wider `isDefinitelyUndefinedExpr`.
The residuals were exactly the type-level `undefined` cases (`var x;`,
`function(){}()`), and they flipped once the intended edit actually landed. The
extraction script asserted its match counts, which is how this surfaced.

Re-measured after the move (nothing assumed — it relocated ~280 lines):

  goal-scope population   43 files:  0 -> 35   (search 15/15, match 11/11,
                                                split 9/9; the 8 `replace`
                                                files stay out of scope)
  all-official population 99 files:  0 -> 47
  regression guard       166 files (baseline standalone `pass` across the six
                                    touched dirs): 166 -> 166, 0 regressed

tsc clean; 20/20 unit tests solo; prettier + biome clean; LOC and oracle-ratchet
gates green against `upstream/main` as base (both false-fail off the fork's
diverged `origin/main`, #4002).

✓
